refactor: Activity에서 Fragment 이동 제어
