+++ app/src/main/java/woowacourse/shopping/presentation/ui/shoppingcart/ShoppingCartActivity.kt
@@ -2,15 +2,21 @@ package woowacourse.shopping.presentation.ui.shoppingcart
 
 import android.content.Context
 import android.content.Intent
+import android.os.Bundle
 import android.view.MenuItem
 import androidx.fragment.app.commit
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityShoppingCartBinding
+import woowacourse.shopping.domain.model.Cart
 import woowacourse.shopping.presentation.base.BaseActivity
+import woowacourse.shopping.presentation.model.CartsWrapper
+import woowacourse.shopping.presentation.ui.payment.PaymentActivity
 import woowacourse.shopping.presentation.ui.shoppingcart.cartselect.CartSelectFragment
+import woowacourse.shopping.presentation.ui.shoppingcart.orderrecommend.OrderRecommendFragment
 
 class ShoppingCartActivity :
-    BaseActivity<ActivityShoppingCartBinding>(R.layout.activity_shopping_cart) {
+    BaseActivity<ActivityShoppingCartBinding>(R.layout.activity_shopping_cart),
+    ShoppingCartNavigateAction {
     override fun initCreateView() {
         initActionBar()
         initFragment()
@@ -29,6 +35,26 @@ class ShoppingCartActivity :
             replace(R.id.fragment_container_view_main, CartSelectFragment(), CartSelectFragment.TAG)
             addToBackStack(CartSelectFragment.TAG)
         }
+    }
+
+    override fun navigateToOrderRecommend(carts: List<Cart>) {
+        val bundle = Bundle()
+        bundle.putSerializable(
+            OrderRecommendFragment.PUT_EXTRA_CART_IDS_KEY,
+            CartsWrapper(carts),
+        )
+
+        val orderRecommendFragment = OrderRecommendFragment()
+        orderRecommendFragment.arguments = bundle
+        supportFragmentManager.commit {
+            replace(R.id.fragment_container_view_main, orderRecommendFragment)
+            addToBackStack(OrderRecommendFragment.TAG)
+        }
+    }
+
+    override fun navigateToPayment(cartsWrapper: CartsWrapper) {
+        val intent = PaymentActivity.getIntent(this, cartsWrapper)
+        startActivity(intent)
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
