+++ app/src/main/java/woowacourse/shopping/presentation/ui/shoppingcart/orderrecommend/OrderRecommendFragment.kt
@@ -9,7 +9,7 @@ import woowacourse.shopping.domain.repository.ShoppingCartRepository
 import woowacourse.shopping.presentation.base.BaseFragment
 import woowacourse.shopping.presentation.base.observeEvent
 import woowacourse.shopping.presentation.model.CartsWrapper
-import woowacourse.shopping.presentation.ui.payment.PaymentActivity
+import woowacourse.shopping.presentation.ui.shoppingcart.ShoppingCartNavigateAction
 import woowacourse.shopping.presentation.ui.shoppingcart.orderrecommend.adapter.RecommendAdapter
 
 class OrderRecommendFragment : BaseFragment<FragmentOrderRecommendBinding>() {
@@ -61,12 +61,7 @@ class OrderRecommendFragment : BaseFragment<FragmentOrderRecommendBinding>() {
         viewModel.navigateAction.observeEvent(viewLifecycleOwner) { navigateAction ->
             when (navigateAction) {
                 is OrderRecommendNavigateAction.NavigateToPayment -> {
-                    val intent =
-                        PaymentActivity.getIntent(
-                            this.requireContext(),
-                            navigateAction.cartsWrapper,
-                        )
-                    requireContext().startActivity(intent)
+                    (activity as? ShoppingCartNavigateAction)?.navigateToPayment(navigateAction.cartsWrapper)
                 }
             }
         }
